Fix a bug in event dispatch.
If an event listener is removed during an event dispatch, it shouldn't receive
the active event. So now in addition to storing a defensive copy, we use a
wrapper object for the event handler to store an `on` boolean. This is set to
false when the listener is removed, preventing that handler from receiving the current event.