Improve performance of re-adding a layer.
Previously, when re-adding a layer to its map, the layer would be removed from
the map only to be immediately added back. This prevents the browser from
efficiently updating the display. Now we do the minimal amount of work, while
still guaranteeing that layers appear in the order they were added.