Add compass control to examples.
I'm not a huge fan of the compass control cluttering up the examples, but on the
other hand, we haven't yet figured out a way to support mousewheel interaction
consistently across browsers. So it's good to have the compass control available
as an obvious way to zoom in. (Dblclick, shift-dblclick, and +/- also work.)