Allow compass pan controls to be hidden.
The panning circle can be hidden by setting `compass.pan("none")`. This option
is now used in the examples to reduce visual clutter while still providing a
reliable alternative to mousewheel zooming. Note that the compass control must
be configured before being added to the map (at least for now).