Limit the size of "world-size" tiles.
If we make the tiles actually world size, then there are precision problems at
high zoom levels as the contents of the tiles are translated by millions of
pixels (67M at zoom level 18). To avoid this, we now limit the maximum tile size
to 2^16, such that the tile contents are always positioned relatively closely to
the center of the map.

In the event that more than one "world-size" tile would be visible (e.g., near
Greenwich), we only show the tile that overlaps the center of the map. Thus,
when the map is zoomed out, we now only show one copy of the data rather than
multiple copies.