Experimental Tipsy integration.
This is a work in progress. It mostly works, but some of the events aren't quite
right so it's possible for the tooltips to get temporarily out of sync with
their targets. A better approach might be to use foreignObject and then allow
the layer to position the tooltips more precisely.