Fixes for adding/removing/restyling controls.
The `arrow`, `drag`, `grid`, `hash`, `wheel` and `dblclick` controls can now be
removed from the map and re-added to a new (or the same) map. The same is true
of the `compass` control, which also supports changing the style properties
(e.g., `radius`) after adding the compass to a map.