Disable GeoJSON tiling if URL is simple.
If the GeoJSON layer's URL is specified as a string and is not a template, i.e.,
it does not contain substitution variables such as {Z} or {B}, we now implicitly
disable tiling as a side-effect. I'm not a huge fan of silent side-effects, but
this is consistent with the behavior for the existing `features` property.

Note that the image layer has a similar check; however, the image layer is still
tiled with simple URLs, but the images skip the request queue since the same
image appears in every tile.