Stash a reference to the (HTML) Image object.
It appears that even if you have a reference to an SVG image object, the browser
can discard the image from the cache. If we store a reference to the HTML image
object on the tile, then the browser is forced to keep the image in memory,
which avoids flickering.