More efficient zoom level transitions.
Previously, zooming in by one caused all the currently-displayed tiles in level
0 to be re-added to the layer's -1 group element. We can avoid these adds (and
similarly on zooming out) if we rename the level 0 group to be the new level -1
group; however, we have to be careful to reorder the levels to maintain the
relative order: -4, -3, -2, -1, +2, +1, 0.

In addition to being more efficient, this will hopefully avoid some image
flickering that occurs when changing zoom levels. It appears that the SVG client
discards images when they are re-added; by avoiding the re-add on both the image
tiles and the level 0 group element, we can hopefully avoid any flicker!

This commit also changes the layer implementation to initialize its container
(and levels) upon construction, rather than waiting until the map is specified.
This simplifies some of the startup logic and paves the way for eliminating the
init hook.